Rename IrregularsDatabase to EnglishIsFunDatabase and change its location Export EnglishIsFunDatabase schema to project directory
